+++ modules/game/include/weapon.h
@@ -25,19 +25,19 @@ private:
 	int targetType;
 	int missileVisualType;
 	int missilesCountPerAttack;
-	double reloadTime;
-	double minDamage;
-	double maxDamage;
-	double missileSpeed;
-	double waitingTime;
-	double missileSize;
-	double noiseRadius;
+	float reloadTime;
+	float minDamage;
+	float maxDamage;
+	float missileSpeed;
+	float waitingTime;
+	float missileSize;
+	float noiseRadius;
 	MISSILE_MANAGER *ptr;
 
 public:
 	WEAPON(const int newId, const int newAttackType, const int newTargetType, const int newMissileVisualType, int newMissilesCountPerAttack,
-		double newWaitingTime, double newMinDamage, double newMaxDamage, double newMissileSpeed,
-		double newReloadTime, int newMaxAmmo, double newMissileSize, int newPenCount, double newNoiseRadius, MISSILE_MANAGER *newMissileManagerPtr) {
+		float newWaitingTime, float newMinDamage, float newMaxDamage, float newMissileSpeed,
+		float newReloadTime, int newMaxAmmo, float newMissileSize, int newPenCount, float newNoiseRadius, MISSILE_MANAGER *newMissileManagerPtr) {
 			id=newId;
 			ammo=maxAmmo=newMaxAmmo;
 			attackType=newAttackType;
@@ -55,25 +55,25 @@ public:
 			penCount=newPenCount;
 	}
 
-	void attack(double fromX, double fromY, double toX, double toY, HBRUSH *b, HPEN *p) {
+	void attack(float fromX, float fromY, float toX, float toY, HBRUSH *b, HPEN *p) {
 		if(maxAmmo!=DEBUGVAR_UNLIMITED_AMMO) ammo--;
 
 		if(attackType==WEAPON_ATTACK_TYPE_PROJECTILES) {
 			for(int i=0;i<missilesCountPerAttack;i++) {
-				double damage;
+				float damage;
 
 				if(minDamage == maxDamage) damage = minDamage;
-				else damage = ((double)(rand()%(int)(maxDamage*100)+(minDamage*100))/100);
+				else damage = ((float)(rand()%(int)(maxDamage*100)+(minDamage*100))/100);
 
-				double angle = abc(fromX, fromY, toX, toY);
+				float angle = abc(fromX, fromY, toX, toY);
 
 				if(DEBUGVAR_ALLOW_SHOT_PRECISION) {
-					double dispersionMax = 10 * DEGTORAD;
+					float dispersionMax = 10 * DEGTORAD;
 
 					if(rand()%2==0) {
-						angle += ((double)(rand()%(int)(dispersionMax*100))/100);
+						angle += ((float)(rand()%(int)(dispersionMax*100))/100);
 					} else {
-						angle -= ((double)(rand()%(int)(dispersionMax*100))/100);
+						angle -= ((float)(rand()%(int)(dispersionMax*100))/100);
 					}
 				}
 
@@ -82,18 +82,18 @@ public:
 		}
 	}
 
-	double getWaitingTime(void) {return waitingTime;}
+	float getWaitingTime(void) {return waitingTime;}
 	int getMissilesCountPerAttack(void) {return missilesCountPerAttack;}
-	double getMissileSpeed(void) {return missileSpeed;}
+	float getMissileSpeed(void) {return missileSpeed;}
 	int getMissileVisualType(void) {return missileVisualType;}
-	double getMissileSize(void) {return missileSize;}
+	float getMissileSize(void) {return missileSize;}
 	int getAmmo(void) {return ammo;}
 	int getMaxAmmo(void) {return maxAmmo;}
-	double getReloadTime(void) {return reloadTime;}
+	float getReloadTime(void) {return reloadTime;}
 	int getMissileTargetType(void) {return targetType;}
 	int getPenCount(void) {return penCount;}
 	int getAttackType(void) {return attackType;}
-	double getWeaponNoiseRadius(void) {return noiseRadius;}
+	float getWeaponNoiseRadius(void) {return noiseRadius;}
 
 	void reloaded(void) {ammo=maxAmmo;}
 };
