+++ modules/game/source/noise.cpp
@@ -1,6 +1,6 @@
 #include "noise.h"
 
-void NOISE::increaseNoiseRadius(double deltaTime) {
+void NOISE::increaseNoiseRadius(float deltaTime) {
 	radiusTemp = radiusTemp + NOISE_SPEED * deltaTime;
 	if(radiusTemp > radiusMax) {
 		markAsUnused();
@@ -9,12 +9,12 @@ void NOISE::increaseNoiseRadius(double deltaTime) {
 
 void NOISE::draw(HDC bhdc) {
 	//SelectObject(bhdc, *ptr);
-	double newX = x + cameraOffsetX;
-	double newY = y + cameraOffsetY;
+	float newX = x + cameraOffsetX;
+	float newY = y + cameraOffsetY;
 	Ellipse(bhdc, newX-radiusTemp, newY-radiusTemp, newX+radiusTemp, newY+radiusTemp);
 }
 
-void NOISE::depose(double X, double Y) {
+void NOISE::depose(float X, float Y) {
 	x = x + X;
 	y = y + Y;
 }
